feature/item Feat: 페이징 추가

+++ src/main/java/com/capstone/rentit/item/service/ItemService.java
@@ -1,10 +1,15 @@
 package com.capstone.rentit.item.service;
 
-import com.capstone.rentit.common.ItemStatusConverter;
+import com.capstone.rentit.common.CommonResponse;
 import com.capstone.rentit.item.domain.Item;
 import com.capstone.rentit.item.dto.*;
+import com.capstone.rentit.item.exception.ItemNotFoundException;
+import com.capstone.rentit.item.exception.ItemUnauthorizedException;
 import com.capstone.rentit.item.repository.ItemRepository;
+import com.capstone.rentit.member.dto.MemberDto;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,7 +31,7 @@ public class ItemService {
                 .itemImg(form.getItemImg())
                 .description(form.getDescription())
                 .categoryId(form.getCategoryId())
-                .status(ItemStatusConverter.fromInteger(form.getStatus()))
+                .status(form.getStatus())
                 .damagedPolicy(form.getDamagedPolicy())
                 .returnPolicy(form.getReturnPolicy())
                 .startDate(form.getStartDate())
@@ -39,26 +44,42 @@ public class ItemService {
     }
 
     @Transactional(readOnly = true)
-    public List<ItemDto> getAllItems(ItemSearchForm searchForm) {
-        List<Item> items = itemRepository.search(searchForm);
-        return items.stream().map(ItemDtoFactory::toDto).collect(Collectors.toList());
+    public Page<ItemDto> getAllItems(ItemSearchForm searchForm, Pageable pageable) {
+        Page<Item> page = itemRepository.search(searchForm, pageable);
+        return page.map(ItemDto::fromEntity);
     }
 
     @Transactional(readOnly = true)
     public ItemDto getItem(Long itemId) {
-        Item item = itemRepository.findById(itemId)
-                .orElseThrow(() -> new RuntimeException("Item not found"));
-        return ItemDtoFactory.toDto(item);
+        Item item = findItem(itemId);
+        return ItemDto.fromEntity(item);
     }
 
-    public void updateItem(Long itemId, ItemUpdateForm form) {
-        Item item = itemRepository.findById(itemId)
-                .orElseThrow(() -> new RuntimeException("Item not found"));
+    public void updateItem(MemberDto loginMember, Long itemId, ItemUpdateForm form) {
+        Item item = findItem(itemId);
+        assertOwner(item, loginMember.getId());
+
         item.updateItem(form);
         itemRepository.save(item);
     }
 
-    public void deleteItem(Long itemId) {
+    public void deleteItem(MemberDto loginMember, Long itemId) {
+        Item item = findItem(itemId);
+        assertOwner(item, loginMember.getId());
+
         itemRepository.deleteById(itemId);
+    }
+
+    private Item findItem(Long itemId) {
+        return itemRepository.findById(itemId)
+                .orElseThrow(() ->
+                        new ItemNotFoundException("존재하지 않는 물품입니다.")
+                );
+    }
+
+    private void assertOwner(Item item, Long userId) {
+        if (!item.getOwnerId().equals(userId)) {
+            throw new ItemUnauthorizedException("자신의 소유 물품이 아닙니다.");
+        }
     }
 }
